Extract Transpilation class into its own file
